requirements.txt
streamlit
Pillow
openai
boto3
hydralit